Move cover to a separate file, rename
